+++ codeception.yml
@@ -1,5 +1,6 @@
 namespace: Tweakwise\Test
 support_namespace: Support
+bootstrap: bootstrap.php
 paths:
     tests: tests
     output: tests/_output
