+++ tests/Unit/Etc/Adminhtml/SystemConfigTest.php
@@ -36,4 +36,3 @@ class SystemConfigTest extends Unit
         );
     }
 }
-
