+++ tests/bootstrap.php
@@ -1,0 +1,20 @@
+<?php
+
+declare(strict_types=1);
+
+spl_autoload_register(static function (string $class): void {
+    $prefix = 'Tweakwise\\Test\\';
+    $baseDir = __DIR__ . '/';
+
+    if (strncmp($prefix, $class, strlen($prefix)) !== 0) {
+        return;
+    }
+
+    $file = $baseDir . str_replace('\\', '/', substr($class, strlen($prefix))) . '.php';
+
+    if (!file_exists($file)) {
+        return;
+    }
+
+    require $file;
+});
